+++ Scripts/Attributes/Editor/DropdownDrawer.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Linq;
 using UnityEditor;
 using UnityEngine;
 using System.Reflection;
+using System.Collections;
 using System.Collections.Generic;
 
 namespace EditorAttributes.Editor
@@ -37,13 +39,17 @@ namespace EditorAttributes.Editor
 			{
 				var memberInfoType = GetMemberInfoType(memberInfo);
 
-				if (memberInfoType.IsArray)
+				if (memberInfoType.IsArray || memberInfoType.GetInterfaces().Contains(typeof(IList)))
 				{
 					var memberInfoValue = GetMemberInfoValue(memberInfo, serializedObject.targetObject);
 
 					if (memberInfoValue is Array array)
 					{
 						foreach (var item in array) stringList.Add(item.ToString());
+					}
+					else if (memberInfoValue is IList list)
+					{
+						foreach (var item in list) stringList.Add(item.ToString());
 					}
 
 					return stringList.ToArray();